feat(lotto): Money클래스 생성으로 원시값 포장

+++ src/main/kotlin/lotto/domain/LottoesRank.kt
@@ -9,9 +9,9 @@ class LottoesRank(
         }
     }
 
-    fun getWinningMoney(): Int {
+    fun getWinningMoney(): Long {
         return value.map { rank ->
             rank.key.prizeMoney * rank.value.size
-        }.sum()
+        }.sum().toLong()
     }
 }
